feat/task-3-home-page: refactor code after review
