+++ app/(main)/_layout.tsx
@@ -1,7 +1,7 @@
 import { Redirect, Stack, useRouter } from 'expo-router';
 
+import { Loader } from '@/components/common/Loader';
 import { useAuth } from '@/providers';
-import { Loader } from '@/shared/Loader';
 import { Ionicons } from '@expo/vector-icons';
 import React from 'react';
 import { TouchableOpacity } from 'react-native';
