+++ app/(main)/components/Amount.tsx
