+++ components/common/Title.tsx
@@ -1,6 +1,5 @@
 import React, { PropsWithChildren } from 'react';
-import { StyleSheet } from 'react-native';
-import { Text, TextProps, Typography } from 'react-native-ui-lib';
+import { Text, TextProps } from 'react-native-ui-lib';
 
 type TitleProps = {
   text: string;
@@ -13,14 +12,8 @@ export const Title = ({
   ...props
 }: PropsWithChildren<TextProps & TitleProps>) => {
   return (
-    <Text style={[styles.title, style]} {...props}>
+    <Text h3 style={style} {...props}>
       {text || children}
     </Text>
   );
 };
-
-const styles = StyleSheet.create({
-  title: {
-    ...Typography.h3,
-  },
-});
