+++ shared/index.ts
@@ -1,8 +1,1 @@
 export * from './tokens';
-export * from './Button';
-export * from './Card';
-export * from './CustomLink';
-export * from './Input';
-export * from './ErrorNotifications';
-export * from './Title';
-export * from './Card';
